--- a/tests/ChunkData/creatable.cpp
+++ b/tests/ChunkData/creatable.cpp
@@ -8,15 +8,20 @@ int main(int argc, char** argv)
 
 	class cMockAllocationPool
 		: public cAllocationPool<cChunkData::sChunkSection>
- 	{
-		virtual cChunkData::sChunkSection * Allocate()
+	{
+		virtual cChunkData::sChunkSection * Allocate() override
 		{
 			return new cChunkData::sChunkSection();
 		}
-		
-		virtual void Free(cChunkData::sChunkSection * a_Ptr)
+
+		virtual void Free(cChunkData::sChunkSection * a_Ptr) override
 		{
 			delete a_Ptr;
+		}
+
+		virtual bool DoIsEqual(const cAllocationPool<cChunkData::sChunkSection> &) const NOEXCEPT override
+		{
+			return false;
 		}
 	} Pool;
 	cChunkData buffer(Pool);