Closes #1244

Initially I was just going to add the `cChunkData` to `cSetChunkData` but profiling revealed that the copying wasn't even the biggest slowdown.  Much more time was being spent in `cChunk::CreateBlockEntities` and `cChunk::WakeUpSimulators` than was in `memcpy` so I've made those significantly faster as well.

Optimisations performed:
 * `cSetChunkData` now stores blocks in a `cChunkData` object
 * `cChunkData` objects can now perform moves even if they are using different pools
 * `cChunk::CreateBlockEntities` now iterates in the correct order and only over present chunk sections
 * Similarly for `cChunk::WakeUpSimulators`
 * `cSetChunkData::CalculateHeightMap` now shortcuts to the highest present chunk section before checking blocks directly

Performance results from timing Cuberite start-up with the world already generated:

| Current Master | This PR      |
|----------------|--------------|
| 1282 ± 76 ms   | 1023 ± 45 ms |

![image](https://user-images.githubusercontent.com/13238737/43050237-cd44921c-8df4-11e8-81e7-a9d21663d64c.png)